发布0.6.2

+++ id-generator-spring-redis/src/main/java/icu/congee/id/generator/distributed/ttsid/TtsId.java
@@ -14,10 +14,10 @@ public class TtsId implements Id {
     private long timestamp;
 
     // 10bit
-    private short threadId;
+    private long threadId;
 
     // 12bit
-    private short sequence;
+    private int sequence;
 
     public static long currentTimestamp() {
         return System.currentTimeMillis();
